network-calculator-old/requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
docopt==0.6.2
fonttools==4.55.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pipreqs==0.4.13
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.6.0
six==1.16.0
ttkbootstrap==1.10.1
urllib3==2.2.3
yarg==0.1.10
